Refactor `LocaleFormatter` logic into `Translation` (without deprecation for now)
