+++ src/workspace.ts
@@ -11,7 +11,9 @@ export {
     SyncStore, useEventStore, useFrameDebouncedStore, useObservedProperty,
     useSyncStore, useSyncStoreWithComparator,
 } from './coreUtils/hooks';
-export { Translation, TranslatedText, useTranslation } from './coreUtils/i18n';
+export {
+    LabelLanguageSelector, TranslatedProperty, TranslatedText, Translation, useTranslation,
+} from './coreUtils/i18n';
 export { KeyedObserver, KeyedSyncStore, useKeyedSyncStore } from './coreUtils/keyedObserver';
 export { WorkerDefinition, defineWorker, useWorker } from './coreUtils/workers';
 export { Debouncer, animateInterval } from './coreUtils/scheduler';
